Add new components and optimise existing ones
This commit adds a number of new components which improve the UI, optimise existing components, and update hooks to work with newly added store modules. A few redundant files have been deleted too. The changes aim to make the application more modular, improve the UI and UX, and make the application more maintainable.